moved json reading logic to a different module so that it can be accessed by calling
